Add group regex scheduling hint to the test partitioner
This commit adds a new optional parameter, group_regex, to TestListingFixture.
The parameter group_regex is used to group the test_ids. By passing a regex
string with the parameter the test partitioner will match the regex to the
test ids and then group by the result. In the case a test id does not match
the regex it will be put in a default group. These groups are then used for
scheduling the partitions. Each test groups is scheduled together so that all 
tests in a group are run together on the same partition. For the purposes of 
scheduling the groups the sum of all test ids in a group is used to schedule 
the whole group in a partition.