Cleanup some leftove files from testrepository
This commit remove some leftover files from the testrepository repo that
were part of the old build system and not needed anymore. The NEWS file
served as the release notes for the project, that will soon be replaced
by reno which we'll start using as we approach the first release.