Start to cleanup docs
This commit makes some doc cleanups to start the process of cleaning
things up and documenting how things are expected to work. Since things
aren't really working yet this is a bit premature, but I needed a break
from all the cleanup work. There is a still a lack of documentation, but
that will be fixed in future patches (likely after things are working).