Just create a repository with testr run if it doens't exist
This commit adds support for testr run to create it's own repository if
one isn't present. If the config file specified is valid but a
repository is not found testr run will now create one.

This is also a revert of the previous commit that tries to create a
repository in the travis install phase, because that doesn't work.