Cleanup .testr.conf in repo
For right now we're still using a testr.conf file to specify how to use
runners. Eventually this will be just always call subunit.run since the
intent with stestr is to just be python. But, before we do that we're
still relying on the config file. This commit fixes the in repo config
file to actually work. (even though stestr run isn't working yet)