Add ouptut_summary method
This commit adds a new method to the output module for printing a
summary view. This matches the current default output of testr run (or
any other command) which just provides some basic info. This will likely
be replaced in future versions, but we need to at least build up feature
parity first.