Add a CLITestResults class and use it for load
This commit adds a CLITestResult class to stestr. This is taken more or
less straight from testrepository although it's modified with a
flattened class heiarchy and removes the ui layer. It's not entirely
clear what function this serves but it is necessary for the subunit
magic to work in load(). I'm not exactly how all the pieces fit into
that puzzle yet.

The load function still doesn't work 100% yet (which is a prereq for
run) it will exit without error now, and properly parse the input stream
but it does not actually write any data to the repository. I'm not
exactly sure where it's going wrong yet...