Move and fix test_listing fixture
This commit removes the fixture subdir which was causing a potential
conflict with the fixtures import (as reported by masayukig) and
creates a separate module for the test_listing fixture. (which will need
to be cleaned up soon) Also some changes are made to the output handling
to fix code that didn't work. It still doesn't execute properly (at
least in a venv) but it's at least trying now.