Switch to stestr for the executable name
While the intent is for stestr to be a drop in replacement for
testrepository having testr be an executable in stestr will cause a file
conflict when coinstalling stestr and testrepository. While we're still
in the bootstrapping period for stestr having the ability to use both
testrepository and stestr is useful so we can compare running with the
2 projects. When we have the basic functionality working for stestr we
can revisit having a backwards compat entrypoint.