Fix stestr load from stdin
This commit fixes the stestr load command when handling a subunit stream
via stdin. It will now handle the input stream properly and write it to
the repository correctly. All that remains for the load command now is
to fix the other 2 call paths, from N file paths on the cli and from an
internal call inside stestr (used for the run path).