Cleanup the manual section in docs a bit
This commit makes some cleanups to the docs for the manual page. Mostly
this is just to reflect the change from the binary name from testr to
stestr. But it also adds a note about stestr load only working with
subunit v2. My earlier assertions that load from a file didn't work was
because I made this mistake (and was only ever passing in a subunit v1
stream). The command works fine if you use subunit v2 streams so this
adds a doc reflecting this.