Nuke the testrepository unit tests
There is no value in keeping the testr unit tests around. They depend
heavily on the previous structure of the repo and won't ever work after
the significant changes made already. It'll just be easier to craft new
tests from scratch. If there are any pieces form the old testr unit
tests that will be useful for the new tests we can always just manually
copy them back in.