Fix testr list command again
When 1bd569892715b68f1c3e6b852d6609f4dd86f8d0 landed it made some
changes to the parallel execution path for setting the concurrency to
run with. This part of that commit inadvertently broke the list command
because it assumed that there was always going to be a concurrency cli
argument defined (which there isn't in the stestr list command). This
commit fixes that bug by treating the lack of a defined option as not
having set one.