Make run command actually run tests
This commit "fixes" the run command so that it actually runs tests. But
it still leaves a bunch of stuff broken. Right now things execute in
parallel but you do not get real time output from the command. The
stream handling is still all messed up. As a result of the same bug
(which I made a note where I added it) it's also not really interleaving
the streams from parallel workers instead appending them serially to the
output stream (after all the tests execute). This is just incremental
progress towards making things work for real.