Fix serialized subunit stream output from stestr run
This commit fixes the issue with stestr run where the subunit output
from the workers is serialized. This was caused by the stream iterator
method calling read() on each stream instead of just returning the
object directly. This read() blocked until the output was finished
meaning that the load() call (and any subunit output) would get the
stream from each worker one at a time. This commit addresses this by
having the iterator yield each stream object directly instead of
reading it first.