Fix --subunit flag on run command
The stestr run command relies on the the load() call to handle it's
output. (the output streams from the workers get passed directly into
load) however there were a couple of places where the passthrough from
the arg passthrough to load() was missed causing the argument to not
work. This commit fixes the oversight so --subunit will work properly.