Fix stestr run on py2
The stream objects are of a different type depending on which version of
python is used. When running with py2 it is a file object, which doesn't
have or require the use of detach() like the buffered reader returned by
py3. This commit addresses this by just having the stream iterator
return the raw file object.