Relicense project under just Apache 2.0
The testrepository project has an interesting dual license where the
user can select to license it either under BSD or Apache 2.0. While
this gives people a bit a flexibility in how they redistribute
testrepository it ends up being a bit confusing. As stestr is a
derivative work I'm choosing to license it under just Apache 2.0 for
simplicity.