Separate out the abstract repository class
The abstract repository class was previously integrated into the
namespace module for stestr.repository (as part of __init__.py). This
was a little confusing and different from the style for the rest of the
project. This commit splits that out into a separate abstract module
to make it more consistent.