Add tests to verify the return code of run command
The return code from the run command is very important because it is
often used as part of CI workflows to verify that a test run passed or
failed. We always want to ensure that a non zero value is returned on
failures and a 0 is only returned if the test run was successful. This
commit adds tests to verify this (and in the process fixes a bug in the
regex filter usage).