Cleanup the config file model
The testrepository config file model was quite involved, mostly due to
it's designed flexibility of allowing any test runner to be used. This
made it difficult to use and understand what's going on. Part of
stestr's design is to clean this up and dedicate itself to just being a
python runner. This commit does just that, it removes all that
configurability from the config file and simplifies it to just 3
options: the test path, the top directory (for discovery), and the group
regex. This makes it much simpler to use.

A future work item would be to add CLI arguments that when used will
override the config. That'll make it much easier for new users to
figure out how this all works.