Add tool to convert testr.conf to stestr.conf
Since stestr isn't strictly backwards compatible with testr user
intervention will be required to start using stestr as a replacement for
testrepository. This commit adds a tools directory and populates it with
a naive script that attempts to automate creating a .stestr.conf file
from the contents of a pre-existing .testr.conf.