Start adding developer documentation
This commit starts the process of having real developer documentation
in stestr. It's far from complete right now, but it is a starting point
for explaining the internal operation of stestr. Hopefully things will
eventually grow to the point where this is a useful guide for new
developers on stestr.