Cleanup the stestr manual
The MANUAL.rst file is basically a more detailed overview of how to use
stestr. Most of this was carried over from testrepository and was out
of date. This commit updates it to try and reflect reality of stestr.
It's probably not 100% yet, but it's still closer.

Also the old testrepository contributors copyright at the bottom of the
doc is removed since it doesn't really apply anymore.