Make unit test attachment fixtures always enabled
The code used in the base test class was copied from other projects and
made the stdout, stderr, and logging attachments enabled optionally by
env variables. However there is no reason to do that since if there is
any output it is best to capture it instead of dropping it. This commit
removes the conditionals and always attaches these.