Add subunit2sql repository type
This commit adds a new repository type for using subunit2sql. It
performs the basic operations, but with a few limitations:

1.There is no attachment support in either directions (into the db or
out of the db).

2. There is no sanity check on whether the db is
created or not. This means that if stestr init has not been run for the
provided repo url it'll fail in an unexpected way when it tries to use
the repository.

3. There is no count() function defined for the repository, meaning the
stestr stats command will not work. This is due to a limitation in the
subunit2sql api as of v1.8.0 where there is no exposed API function to
get the number of runs.

4. When writing to the repository the entire contents are stored in
memory until the run is complete. This means 2 things stestr will have
at least 2 complete copies of the stream in memory while processing it.
Also if there is a power failure, hard shutoff, or the process is
killed while running tests there will be no partial stream stored in
the repository.

But, other than those issues it works fine. This is just a starting
point and won't be enabled by default for some time.