Workaround stestr last issue with sql repository type
When the sql repository type is used the stestr last command would
encounter an issue trying to calculate the id for the previous run.
The file repository type used an interger so to get the previous run id
it would just subtract 1. However the sql repository type uses a uuid
so this does not work. This commit works around this by not using the
previous run data for now. Eventually we'll add a function to the
repository abstract class to get the previous run id.