Cleanup test_listing_fixture and add docs
This commit cleans up some of the cruft from the test_listing_fixture
(there is still a bit more in there) that was leftover from
testrepository. A bunch of unused parameters were removed as was the
per-instance command code which is never used in stestr (since it
doesn't support running remote workers). As part of this cleanup API
docs were added, including some prose around what it is and how it's
constructed. This is one of the more confusing classes so I though it
was good to try and explain it's purpose.