Fix _Subunit2SqlRun.get_test() method
Previously the get_test() method in the _Subunit2SqlRun class was based
heavily on the version in the file repository type. However the mistake
there was while get_subunit_stream() returned a v2 stream, get_test was
assuming a v1 stream (which is why the get_subunit_stream() call is
commented out there) However in the sql repository type only v2 is ever
used. So using the calls assuming a v1 stream caused the v2 stream
generated by _Subunit2SqlRun.get_subunit_stream() to just passthrough
the subunit library calls() to stdout. (because that's the default
behavior in subunit)

This commit fixes this issue by using ByteStreamToStreamResult(),
which is the correct class, to convert the subunit v2 stream into
a Result object. This fixes both the subunit v2 leaking to stdout as
well as any of the operations which involved using historical data
from the repository.