Add repository presence check to sql RepositoryFactory
This commit adds a check to the sql repository type's RepositoryFactory.
The check just runs the get_ids_for_all_tests() subunit2sql db api
function. This function was chosen because it is read only and will not
raise an exception in the presence of an empty database. (but it still
requires the schema) If the function does not raise an operation error
from sqlalchemy that means stestr was able to communicate to the
database and the subunit2sql was present. But, if an operation error is
raised that means either stestr can't connect to the DB or the schema
isn't present (or up to date). This should be sufficient for the check
during the RepositoryFactory's open() function.