Ensure that we run subprocess with the same version of python
Right now it is possible to run stestr with one version of python and
have it's subprocesses (which are launched for test listing and running
tests) executed with a different version. This would be
caused by relying on the version of python being consistent with what
is set in /usr/bin/python (or equivalent on other OSes). To ensure that
all of stestr is run under the same version of python we export the env
variable PYTHON (which is respected by the subprocess command) to be the
same version of python stestr was launched with, assuming it isn't
already set.