Strip attrs from test_ids before storing time data
The attrs attached to a test are metadata and not actually part of the
test_id. They are appended to the test_id in testtools/subunit so they
are easily filterable, however the attributes on a test can and do
change over time without any change to the test. Since stestr only runs
with subunit.run we can assume a standardized way of appending the attrs
and that lets us strip them out for tracking purposes. This commit does
just that and ensures that we track 2 tests with the same id as the same
regardless of what the attrs are set to.