Add python 3.6 support to stestr
With python 3.6 officially released (and now the default python on
archlinux) add support for it to stestr. This means adding it to the
default tox env, travis config, and a trove classifier in the setup.cfg
At the same time this commit removes some stale trove classifiers for
versions of python that aren't tested.