Strip attrs from test ids in _get_test_times lookup
The subunit2sql code being used to populate the tests table (which is
where the running mean of test times is queried from) strip the attrs
from the test_ids. However the lookup side in the Repository code was
still using a test_ids with attrs. This commit fixes this so lookups
will actually match.