Enable real time DB writes for sql repository
Previously the sql repository type would wait for the subunit stream to
complete before it would start writing the subunit2sql DB. This would
potentially cause an issue if the stream was aborted for any reason
during the run there would be no partial data stored anywhere.

This commit solves this by writing data to the DB as it comes in over
the subunit. An empty run is created at startTestRun in the result
stream, each test_run is added to the DB as it's finished, and the
run_time is added to the run at the end during stopTestRun. This solves
the issue, the only cost is there is more DB activity required to
accomplish this. Since the run also needs to be updated for each test
(to keep the counters up to date).