Preserve full test_id in _get_test_times for file repository
This commit ensures that the test_id without attrs is used for the
timing lookup but when we return the results dict we need to return the
test_id with attrs. If there are no attrs in the test_ids in the results
dict the scheduler will not be able to match any tests with timing data.