Port test selection logic from ostestr
The os-testr project was started to be a bandaid for openstack around
using testr. Mostly to add on to it's CLI and work around some of the
bugs. Part of that was improved test selection logic. This commit ports
the underlying functions to stestr to enable the same type selection.