Add section to README about migrating from testr
There is a helper script in tools/ to migrate a project using
testrepository but no references to it exist in any of the
documentation. In order to make it easier for people to migrate from
testr to stestr, this commit adds a section to the readme on using that
script.