Add --no-discover option to stestr run
This commit adds an option to stestr run, --no-discover/-n, to bypass
test discovery and just directly run a single python callable. This
will make it easier (and much faster) for people iterating on a single
test to just run it directly.

An implementation note here is that subunit.run is called with
subprocess instead of manually calling the subunit.run module. This is
for 2 reasons (which are related) first is that all the glue code in
stestr itself (mostly in stestr.commands.load) is expecting this, since
it is how we execute subunit.run in the case with discovery. Changing
this assumption would mean having to add an additional code path to
stestr load for an internal stream but not from a subprocess.

The second is that the subunit.run api makes it difficult to handle the
streams in a serialized manner. The subunit.run module is basically just
a wrapper for testtools.run which is itself just a wrapper for
unittest.main. In that call path there are a lot of assumptions made
about the input and output types and the control available on those
objects. (for example an io.BytesIO for the output stream does not work,
but real files do) It makes it difficult to have tests run directly in
python. The closest I got to making this work inside a single python
interpreter was to use os.pipe() for the stream and launch a
threading.Thread() to call subunit.run and write the output into
the pipe and then pass the reader side of the pipe to
stestr.commands.load. However even doing this caused an IOError on exit
because something in the subunit.run call path was deleting the file it
was passed.