Add CLI option to stestr run to randomize test order
This commit adds a new option to randomize the test order. This is
useful because in certain situations you want to introduce a bit
of entropy into the execution order. When this option is used the
same scheduler test partitioning strategy is used, but before the
list of tests is passed to subunit.run for execution the list is
shuffled.