Use the same dbm version across python versions
This commit updates the dbm import so we end up using a consistent dbm
version across python versions. Previously if you ran python2 first it
would blow up when you tried to read the db with python3. This was
because of a backwards incompatible change in python's dbm modules. By
using six.moves and gnu dbm always we can avoid this issue.