Workaround potential backwards incompat with gnudbm usage
The earlier commit, 9869bfe48ed87d9abf3f3397c46a9d7f8dc60031, switched
the dbm usage of the file repository to always use gnu dbm. However this
is a potentially backwards incompatible change. If someone has a
repository that was initialized with a dbm file other than gnu dbm.
This commit addresses this by catching the error raised by gdbm in that
case and recreating the times database with gnudbm. The data will be
lost, but since only 1 previous timing record is kept and it's only used
for scheduler optimization this is not an issue.