Fix scheduler.partition_tests randomize option
This commit fixes the scheduler.partition_tests randomize option. I
noticed that 'out_parts.append(set(temp_part))' returns
 [{'c_test', 'd_test'}, {'b_test', 'a_test'}]
but 'partitions' is
 [['c_test', 'd_test'], ['b_test', 'a_test']]
(in python3).

However, we can't test it because it can be same as randomize=False in a
small chance. It might be rare, however unstable tests can be confusing.