Move sql requirements to setuptools extras
The sql repository type is still experimental and not really ready for
general use yet. It's requirements are also fairly bulky and pull a lot
in. To avoid this until it's ready for general use this commit moves the
requirement for using the sql repository type, subunit2sql, into
setuptools extras. This makes installing the requirements opt-in. The
failure caused by not having this installed is handled gracefully with a
message explaining how to install the missing requirements via pip. The
docs are also updated to reflect this.