Print nice error message when test_path isn't set
If you are running stestr and somehow end up in an
unexpected directory (for example when running within
an IDE) the config file may not be found, and the
test_path variable in stestr/config_file.py will
not be set, resulting in an exception. This patch
will check for a non-None value for test_path and
print a (hopefully) useful error message when
test_path isn't set.