Use dbm.dumb rather than gdbm
This commit uses the 'dumb' dbm module from stdlib which is a pure
python implementation of dbm. It works on both python 2 and python3
across all the operating systems. The docs warn against using it because
of a potential performance issue, but testing has shown this isn't a
problem for stestr.