Add support to stestr load to append to an existing entry
This conmmit adds suport to the stestr load command and the underlying
repository API to append a stream into an existing entry in the DB. In
the repository API this is accomplished by adding a new optional
argument on the _get_inserter Abstract repository class to specify a
run_id. When specified the inserter returned will be for that already
existing entry in the repository. The stestr load command just leverages
this new mechanism by taking an id field on the cli and passing it to
the invocation of get_inserter.