Cleanup output from test listing fixture
The test listing fixture had some leftover output straight from testr
which can be confusing. First is it prints the subprocess commands that
were being run, which is unecessary and confusing if you don't know what
is going in on internally. Secondly when it encounters an import it
raises an unhandled exception to crash. This commit fixes this by
printing out a more detailed view of what happened with an explanation.
Also a leftover failure condition from testrepository was removed
because it can never be tripped in stestr.