Add file repository skips for windows
Some of the unit tests for the repository are testing things which are
specific to a unix system. For example that paths which start with ~ are
expanded to a user's home directory. Since these concepts don't exist on
windows lets just skip these tests so we can enable a windows ci.