Add a concurrency option to the manual scheduler
This commit adds a new option in the manual scheduler yaml syntax. There
are certain situations where you want to treat a subset of tests being
manually scheduled like the default case with the standard test
scheduling mechanisms. To enable doing this a 'concurrency' field can be
used on a worker, and tests on those workers will be scheduled like
normal across the number of workers specified.