Rename the test_listing_fixture to test_processor
The name test_listing_fixture was a bit confusing. It was a hold over
from test_repository that was never renamed. Even in testrepository the
function of the module is much more than simply listing tests. It's
where all the test workers are launched. While if you look at the
internals of the module the life of the fixture is around the temporary
test list files it creates. But, this name is confusing for people not
super familiar with the internals.

To try and clear up any potential confusion here this commit renames it
to the more generic 'test_processor' (and the Fixture class inside).
This hopefully makes it more clear that the role of the module is around
processing the tests, and more than just listing things.

Closes Issue #21