Write subunit2sql db manage output to local stdout
Previously the subunit2sql-db-manage command is called via subprocess
call() function. This mean that the stdout and stderr from that command
was being missed by the stdout and stderr fixtures because it wasn't
originating from the parent process. This commit fixes this by storing
the stdout and stderr in a subprocess.PIPE and writing it to the parent
process's stdout and stderr at the end of the execution.