Copy subunit-trace and colorizer from os-testr
This commit copies over the subunit-trace and colorizer (it's only in
repo dependency) from the os-testr project. [1] These modules will be
used as the new default output filter from stestr, but to avoid a
potential circular dependency (the long term plan is to switch os-testr
to use stestr underneath) we need the subunit-trace modules to live
locally.

[1] http://git.openstack.org/cgit/openstack/os-testr