Fix stestr --test-path argument for stestr run and list
This commit fixes a small bug in the recent python api change that
causes the --test-path argument to not work. The if condition for
determining if a test_path is not set in either the config file or the
cli was mistakenly assuming if it was set on the cli was the same as not
being set. This commit fixes it by explictly checking if it's not set.