Add stdout argument to command functions
This commit adds a stdout argument to the command functions. If you're
calling a command function from python there are several use cases where
you don't want the output to unconditionally go to stdout. To enable
such use cases this commit adds a stdout arg to all the functions that
write to stdout. By default sys.stdout will be used as before, but any
file object can be used instead.