Handle stestr init with existing path
Previously if you ran stestr init with a file type repo on a path where
the repository already created you would get a stack trace. That's
because the file layer will go to do a mkdir and if the path already
exists it will raise an exception which bubbles up to the cli layer.
This commit just catches the FileExistsError and prints a message
instead.