Add unit tests for return code from functions
This commit adds unit tests calling the command functions from the
python interface to verify their return codes. As things start to
consume the python interfaces it will be good to point to tests that
validate the return codes and functionality are the same as running from
the CLI. The tests added here are just a mirror of our cli return code
tests but instead of using subprocess they call the python api.

The python-subunit minimum version is bumped as it contains fixes
necessary for the tests to work.