Create a Code of Conduct for the project
Github was bugging me about adding a code of conduct to the repository
and I actually think it's a good idea. Despite having very few
contributors to this project and a small community around it. If it
ever grows or there is a problem having a code of conduct written down
will at least set the expectations just in case.