Use context managers for file open()
This commit tries to use context managers for file open(). In the
previous commit, it already has been done as a bug fix. However, these
also can be fixed to close files. In some cases, the original code are
not harmful, but I changed it to keep the code clean.