Fix the stestr run --analyze-isolation flag
The analyze-isolation command was never update after the stestr repo was
forked from testrepository. This code path was still making a bunch of
calls assuming the repo was still laid out like testrepository. This
commit fixes all of these and refactors the test biscet code to live in
it's own class. This also adds tests for this logic to ensure it works.

Fixes issue #85