Add a --abbreviate flag to stestr run and load
This commit adds a new flag to the stestr run and load comands for
printing abbreviated output. By default the full subunit-trace output is
used, but in some cases this is overly verbose. To enable a tradeoff
between the full verbosity and the blank output during execution this
flag will print a single character as each test executes.