Fix if logic on output types in stestr load
The if logic in stestr load was incorrect and when subunit_out was
specified it would set the right variables, but then override them later
in the code path. This commit fixes it so that when subunit_out is used
those settings hold true until the actual stream is run. In the process
another bug was fixed when subunit_out and pretty_out were both set the
subunit_trace summary was still printed because of a similar if logic
issue.

Closes Issue #101