Respect all selection mechanisms for discovery check
When test_processor is doing it's initial test_id list generation it has
to determine whether or not discovery is necessary. Whenever a dynamic
test selection mechanism (regex, black regex, whitelist, or blacklist)
is used running discovery is necessary if test_ids aren't supplied.
However, the check for this only understood the normal regex and wasn't
checking if any of the other selection mechanisms were in use. This
commit fixes this issue by checking if any of the selection mechanisms
are in use instead of just a regex.

Fixes issue #103