Add test_return_codes check for valid subunit output
This commit adds a check to the test_return_codes module to ensure that
valid subunit is being emitted when we say the command expects subunit
output. This check is accomplished by processing the output stream and
converting it to a list of dicts. Then we check that at least 1 test was
present in the output stream (which should be the case for all the
tests) to ensure it actually is processing the subunit.

This is a follow on unit test for #102