Fix stestr load from stdin
This commit fixes a bug in stestr load when passing in a subunit stream
from stdin. It was previously trying to read the stream form the stdin
buffer, but that doesn't always (and in most cases) exist. It turns out
there was no need to do this, so this commit just fixes that by using
the stdin object itself.

Fixes #109