Fetch the data from last run earlier
The run command, before bisect is used, tries to get rid of tests
that always fail - to do so it generates new runs, which loses
the reference to the last one before bisect was requested.

To fix this, fetch the last run earlier in the run code.
On bisect side, run all candidate tests, not only failing ones.
And drop an extra comma which broke things.

This version of bisect only find conflicting tests within a
worker since the bisect code relies on the ordering of tests in
the database for a worker.