Add functional test for analyze-isolation flag
The --analyze-isolation stestr run flag is fairly complex and unit tests
don't really go far enough to test it's operation. So, in addition to the
unit tests it is useful to have some functional test coverage for the
flag to verify end to end operation. This commit starts the file for
doing this with a single functional test. This gives us the groundwork
to expand the testing in the future as we find more corner cases in it's
operation.