Remove 32bit windows testing CI config
The appveyor windows testing is much slower than travis, mostly because
the different jobs are run serially. Submitted patches are often left
waiting for several minutes while it finishes. We can reduce this wait
time by limiting the number of windows jobs we run. This commit does
that by removing the 32 bit windows jobs. We were previously running
each python version twice once on 64 bit windows and again on 32bit.
This is a bit unecessary as this is very unlikely to ever cause an
issue. It was also a waste of test resources.