Pass output to output.make_result() in stestr load
The stestr load command was not passing the stdout arg to the
make_result call, which is what we use to generate the subunit output.
This commit fixes this oversight and passing stdout to the function to
make sure we write the subunit output to the specified file instead of
unconditionally using sys.stdout.

Fixes Issue #115