Implement quiet option for run and load commands
This commit makes run and load commands use the existing quiet option
and implements it. We don't need outputs of run and load commands
occasionally. So, this patch adds the function to suppress the outputs.

It might be better to move the quiet (or silent) option to load and
run commands because the option doesn't work in the other commands now.
However, if we move the option to the subcommands, it causes a backward
incompatibility.

Fixes issue #121