Handle zero and negative times in stestr slowest
To calculate a starting point for the column width for printing the
slowest tests table the first result was pulled from the list of test
times and the number of digits was counted. However, in the case of a
0 or negative number this doesn't work because taking a base 10 log of
that will fail. This commit address this corner case by using the first
value > 0 and if none can be found defaulting to 1 digit.