Fix warning message about missing worker tags in subunit-trace
This commit fixes a small oversight in a warning message from
subunit-trace where a worker tag is not found in the output stream. This
originally was added to subunit-trace to expose a race in testrepository.
However attributing this error to testr here has 2 issues, first that
particular bug has been fixed in testrepository and the error can't ever
come from  testr in the stestr embedded copy of subunit-trace. This
attepmts to make the warning message a bit clearer if anyone ever
encounters it from stestr.