Add support for a user config file
This commit adds a new feature, the user config file. This enables users
to write a global config file for UI preferences on certain commands
that will be the new default for that option. For example, if you prefer
the default output on the 'stestr run' command to be with the
--no-subunit-trace flag set. You can just write a user config file to
set this output. Then for all stestr run invocations the default output
will be with --no-subunit-trace set.